[#8] The activeFeatures is just an array, not an Object
We don't need a key/value pair. The key is the same for all (resourceKeyField),
we're just adding values.

This might make it slower, because we have to search in an unordered array
every time someone clicks on a region (O(n) complexity), against simply
checking the existence of a key in an Object (O(1) complexity). But, as there
probably won't be many active filters at once, it shouldn't make much
difference.