[#8] Activate geojson's features when loading the map with filters already set
I tried isolating this routing/activating logic as much as I could. Actually,
they should be in a separate file, with most parts being reused between all
resource views that understand those filters, but we're not there yet.

The name "router" is a bit misleading, as it's not only routing, but styling
the features, but I couldn't think of anything better.