[#8] Hack to disable clicks and change pointer when we're not in a dashboard
I had to do this because when we're not inside a dashboard, it makes no sense
to click and set up filters. I had to hack my way through because, AFAIK,
there's no way to disable click events while keeping hover events in LeafletJS.
I created a feature request in their repository to see if we can find a better
way (https://github.com/Leaflet/Leaflet/issues/2499).

Meanwhile, the biggest disadvantage that I see with this approach is that we're
coupling the map with the dashboard, by checking if we're inside a
.dashboard-grid. It's bad because maybe the user would like to have a
"navigatable map" in her blog post, where she manually added a few charts.

A solution for that could be creating a new resource preview: navigatable map.
So she could choose what she wants to do. It will still be bad, because if a
CKAN user sees that map in the resource previews list, he'll be able to click
on the features and change the URL with no effect, but might be a bit better
than what it is now.