[#9] Now you can select which URL we'll redirect to when clicking on regions
By default, clicking on a region when inside a dashboard will simply set the
filters in that same page. This was working already.

Now, if the user sets a redirectToUrl in the map's edit form, clicking on a
region will redirect you to whatever URL she configured, with the filters
already defined.

When working on it, I found a problem: if redirectToUrl is None or '', setting
it on the HTML will turn:

  data-module-redirect-to-url = "{{ resource_view.redirect_to_url }}"

into

  data-module-redirect-to-url

with no value defined. In the JavaScript side, that's interpreted as a boolean
toggle. So we end up with options.redirectToUrl == true. The best workaround
I've found was testing in JS if it's true and, if so, consider that it's blank
('').

The code looks odd, but I was unable to find a better option.