[#9] Disallow clicking only when we're in our own resource view page
If the map is being shown outside its own resource view page (for example, in a
dashboard, or embedded in a page, etc.), it is interactive.

Again, the way I'm detecting if we're in the resource view page or not is very
ugly, but I couldn't figure out a better one.