[#13] Begin refactor extracting choroplethmap creation from CKAN's module
The final aim is having a reusable choropleth map which knows nothing about
CKAN's resources. We might not achieve that, but at least I want to be able to
have a choroplethmap CKAN module, and a regular map (or navigationalmap)
module, one which creates a choropleth, and other that creates a regular map,
reusing as much code as possible.