[#13] Extract map's ckan module creation to a separate file
First of all, the visualizations shouldn't know about CKAN. I also removed
quite a bit of code duplication doing this. I had to add some magic to do this
(by using the same initializer, just changing the type of map to be called),
but I think it's better than it was before.