[#16] Don't add repeated filters when activating regions
This works even if the repeated filters come from one of the other fields the
user configured. For example, say that we're plotting a map with the states in
the USA, and I add ```country``` to the fields as well. So, when clicking on a
state, we'll p filters for it AND for the country, as in:

```?state:USA|state:California```

The problem is that when I click on another state, it'll add the country
n,
as in:

```?state:USA|state:California|state:USA|state:Florida```

We shouldn't be able to duplicate filters. What we want is:

```?state:USA|state:California|state:Florida```