[#3] Add selectbox to pick the GeoJSON resource
Even though we show the resource name, we're saving just its URL. This means
that if the user re-uploads the resource, this will point to the old URL and
will have to be updated as well.

I haven't saved the resource id itself for performance reasons. If we didn't
had the URL, we'd have to get it before loading the map. Every time. We might
want to do this way in the future, anyway, as the performance might be
acceptable when it's cached.